Removed FeatureRDDFunctions class.
* Removed `FeatureRDDFunctions` class and moved all methods into `FeatureRDD`.
  Renamed test suite to match, and added Java helper functions.
* Moved all `adam-*/src/test/resources/features/*` files into their respective
  `resources` directory. This is necessary due to an issue when using
  `copyResources` with subdirectories.
* Removed `ignore` on test in `org.bdgenomics.adam.cli.Features2ADAMSuite`.
