+++ adam-cli/src/test/resources/NA12878.sam
