+++ adam-cli/src/test/resources/chr5.phyloP46way.trunc.wigFix
