+++ adam-cli/src/test/resources/gencode.v7.annotation.trunc10.bed
