+++ adam-core/src/test/resources/SeqCap_EZ_Exome_v3.hg19.interval_list
