+++ adam-core/src/test/resources/dvl1.200.gtf
