+++ adam-core/src/test/resources/gencode.chr20.transcript_names.head10.txt
