+++ adam-core/src/test/resources/gencode.v19.annotation.chr20.250k.gtf
