+++ adam-core/src/test/resources/wgEncodeOpenChromDnaseGm19238Pk.trunc10.narrowPeak
